requirements.txt
aiogram==3.13.0
python-dotenv==1.0.1
asyncpg==0.29.0
django==5.1.1
psycopg2==2.9.9
fastapi==0.115.2
uvicorn==0.31.1
python-multipart==0.0.12
pandas==2.2.3
matplotlib==3.9.2
xlsxwriter==3.2.0